Refactor GPIO handling
Moves GPIO from main.c to d39_gpio
