+++ CMakeLists.txt
@@ -7,4 +7,5 @@ find_package(Zephyr)
 project(d39_fuel_flow)
 
 # Add application code
-target_sources(app PRIVATE src/main.c src/nmea.c)
+target_sources(app PRIVATE src/main.c src/nmea.c src/d39_gpio/d39_gpio.c)
+target_include_directories(app PRIVATE src)
