+++ src/d39_gpio/d39_gpio.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "stdint.h"
+
+int d39_gpio_init();
+uint32_t d39_gpio_get_and_reset_ticks();
